initalise new zealot campaigns with empty zealot progress object
